# PERF: added no exception versions of '_string_to_dts' and 'parse_iso_8601_datetime' functions