--- a/pandas/_libs/tslib.pyx
+++ b/pandas/_libs/tslib.pyx
@@ -584,7 +584,7 @@ cpdef array_to_datetime(ndarray[object] values, str errors='raise',
                         string_to_dts_failed = _string_to_dts_noexc(
                             val, &dts, &out_local,
                             &out_tzoffset
-                        ) != 0
+                        )
                         if string_to_dts_failed:
                             # An error at this point is a _parsing_ error
                             # specifically _not_ OutOfBoundsDatetime