--- a/pandas/_libs/tslibs/conversion.pyx
+++ b/pandas/_libs/tslibs/conversion.pyx
@@ -467,10 +467,10 @@ cdef _TSObject convert_str_to_tsobject(object ts, object tz, object unit,
                     else:
                         # Keep the converter same as PyDateTime's
                         obj = convert_to_tsobject(obj.value, obj.tzinfo,
-                                                None, 0, 0)
+                                                  None, 0, 0)
                         dt = datetime(obj.dts.year, obj.dts.month, obj.dts.day,
-                                    obj.dts.hour, obj.dts.min, obj.dts.sec,
-                                    obj.dts.us, obj.tzinfo)
+                                      obj.dts.hour, obj.dts.min, obj.dts.sec,
+                                      obj.dts.us, obj.tzinfo)
                         obj = convert_datetime_to_tsobject(
                             dt, tz, nanos=obj.dts.ps // 1000)
                         return obj