- [x] closes #N/A
- [x] tests added / passed
- [x] passes `git diff upstream/master -u -- "*.py" | flake8 --diff`
- [ ] whatsnew entry

`asv continuous -f 1.05 origin/master HEAD -b ^io.csv -b ^timeseries -a warmup_time=1 -a sample_time=1`:

master| patch| ratio| test name
-|-|-|-
98.4±3ms|         76.2±2ms|     0.77|  timeseries.ToDatetimeFormatQuarters.time_infer_quarter
15.2±0.6ms|      9.75±0.04ms|     0.64|  io.csv.ReadCSVParseSpecialDate.time_read_special_date('mdY')
37.1±2ms|       22.4±0.1ms|     0.60|  io.csv.ReadCSVParseSpecialDate.time_read_special_date('mY')