@anmyachev 2019-04-27
> So the speedup here is purely replacing the "except? -1" clause in the cdef with manual error checking right?

@WillAyd you're right.