@WillAyd 2019-04-27
OK thanks. Do you know if the question mark in that statement has a performance hit? 

Not against this change just surprised to see it make that big of a difference. Certainly being able to leverage the built-in Cython functionality would be simpler and preferable if removing the question mark gets us closer (hence my question)