@anmyachev 2019-04-27
> OK thanks. Do you know if the question mark in that statement has a performance hit?

The performance degradation from this is unimportant because most likely this is one additional `PyErr_Occurred()` call, I think.

> Not against this change just surprised to see it make that big of a difference. Certainly being able to leverage the built-in Cython functionality would be simpler and preferable if removing the question mark gets us closer (hence my question)

The main performance increase is that we stop throwing a `ValueError` exception (this is too expensive for python). So removing the question mark will not allow performance gains and will most likely change the logic of date parsing.