@jbrockmendel 2019-04-27
> The main performance increase is that we stop throwing a ValueError exception (this is too expensive for python)

@anmyachev any idea how many more formats/functions you plan to optimize like this?  In the past we've been trying to get code _out_ of the C files and to stay python-idiomatic where possible.  It looks like you're making an effort to keep the non-idiomatic code contained to np_datetime, which is good.

Could the raising of ValueError be made more efficient upstream in cython?  Just spitballing: if the return-code based version is really equivalent to the raising-based version, it seems like cython should be able to figure that out.