@anmyachev 2019-04-27
> @anmyachev any idea how many more formats/functions you plan to optimize like this? In the past we've been trying to get code _out_ of the C files and to stay python-idiomatic where possible. It looks like you're making an effort to keep the non-idiomatic code contained to np_datetime, which is good.

At the moment, nothing more in this way. This function was "hot" when profiling, and the main slowdown was just in throwing an exception. Other functions with exactly the same problem have not been identified yet.

> Could the raising of ValueError be made more efficient upstream in cython? Just spitballing: if the return-code based version is really equivalent to the raising-based version, it seems like cython should be able to figure that out.

Probably, but the main problem is the speed of exceptions in the python itself, and not how сython generates it.

In the first case, we need the function to throw an exception, and in the other it does not, depending on our choice(for this we use `want_exc` argument). I cannot achieve this purely with the construction of a cython.

If the existence of functions "_string_to_dts_noexc" and "_string_to_dts" is not very good for you, then I can try to make one function, but with an additional parameter (to throw an exception or not) while leaving with the instruction "except? -1". What do you say?