@jbrockmendel 2019-04-27
> If the existence of functions "_string_to_dts_noexc" and "_string_to_dts" is not very good for you [...]

Not a problem at all.  Why would we ever use the slow version?

> Probably, but the main problem is the speed of exceptions in the python itself, and not how сython generates it.

_string_to_dts is `cdef`, so it can only be called from C/cython.  The idea was that for these functions, cython could make the substitution and only ever do the raising (slow) in a python-exposed function.