@vnlitvin 2019-04-28
> The idea was that for these functions, cython could make the substitution and only ever do the raising (slow) in a python-exposed function.

This is possible if desired, we'd just have to make `parse_iso8601` return different error codes to specify correct exception messages.

The exceptions themselves are not _that_ slow, but in this particular case they were taking way more time than `parse_iso8601` function itself (because the function on usual data very quickly either parses the string or figures out it's not a ISO8601 datetime - nothing of this requires new allocations), and raising an exception requires allocating a few objects (at least the exception and the traceback objects) and unwinding the stack to store in the traceback. If the only thing we care (as in this particular case) is whether exception happened _at all_ (and we don't care about message or stacktrace), then we should be fine using boolean "error happened" flag.

As for Cython being smart there - I think it might be possible to make a keyword or something that would tell Cython to generate some error-code-based error checking and hide it pretending it's exceptions (like `@nogil` pretends it's still Pythonic but releases GIL).