@anmyachev 2019-04-28
> this is adding a lot of complexity for a relativily small gain. If you want to remove the exception handling, then just do it and force a check everytime it is used. having 2 function is way more complex code.

@jreback "noexc" version of functions removed