@anmyachev 2019-04-29
@jreback I simplified changes.