@anmyachev 2019-05-03
@jreback @WillAyd all CI builds (include `macOS py35_macos`) actually completed successfully.