@TomAugspurger 2019-05-10
This seems to have slowed down `timestamp.TimestampConstruction.time_parse_iso8601_tz` http://pandas.pydata.org/speed/pandas/index.html#timestamp.TimestampConstruction.time_parse_iso8601_tz?commits=71207255f32af73cd701fe4ecf56a643c925a9e1-2e4e0b9058cf9f5a3ca81d2fb405fcd3b9317951